documentation file
I started a documentation file that we can look back on for the readme, and will also serve as a good track of our progress through the project. this file is structured as date/ntime/nauthor/n for each section and is to be commited on each addition to the file seperately from the added code and to show in its commit message a good review onto it. Best wishes on this project
exit