added operations
i described an efficient way to describe the combnartions that describe the operations with full details in the documentation file. I also started the MakeFile so that we will add onto it gradually whatever files and folders we have. The makefile also compiles the makefile of libft so we can benefit from that. I will add ft_printf later to use it as well, so we will compile it similarly. I also researched all algorithms we will be needing of O(n squared) complexity that i will soon add to the doc. The next update should include input parsing/ compiling ft_printf/ error handling/ transforming from string input to linked lists (by some miracle :)), that is all for nowgit add .