bugs all over the place :(
i cannot think anymore there are very obvious bugs from the single simplest test i tries which is the basic input 1 3 2, the issue can be as simple as an uninitialized variable or a whole logical issue i havent thought about it, i have updated the join sep function so that if the 1st string was null we just set s1 = strdup s2 and return so we dont have the issue of having \n at the start of the result being returned. i also tried a test main function but that is definitely not the one we will settle on. i hope the issues will be minor and i will fix them in the next updates. gn